Add expense summary to dashboard page
You can now add/remove expenses and have the app summarize the totals for the number of expenses you have for you.

selectExpenses - used to grab the # of expenses the user wishes to see the summary of.

selectExpensesTotal - used to calculate the summary of the expenses.

JEST: snapshot testing applied to ExpensesSummary.js