D3 + Older react version + Babel = doesn't work.
We can't use hooks since we did this on an older versioned react. We can't use subclasses because Babel doesn't like it and so the D3 implementation of the pie chart throws errors regarding layout rendering and babels build doesn't properly handle D3.

May look into building a project that's more up to date with react so I can use hooks.